+++ src/hapi_schema/db_national_risk.py
@@ -6,13 +6,13 @@ from sqlalchemy import (
     Boolean,
     CheckConstraint,
     DateTime,
+    Enum,
     Float,
     ForeignKey,
     Integer,
     String,
     Text,
     select,
-    text,
 )
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -22,31 +22,45 @@ from hapi_schema.db_dataset import DBDataset
 from hapi_schema.db_location import DBLocation
 from hapi_schema.db_resource import DBResource
 from hapi_schema.utils.base import Base
+from hapi_schema.utils.constraints import (
+    general_risk_constraint,
+    reference_period_constraint,
+)
+from hapi_schema.utils.enums import RiskClass
 from hapi_schema.utils.view_params import ViewParams
 
 
 class DBNationalRisk(Base):
     __tablename__ = "national_risk"
     __table_args__ = (
+        general_risk_constraint("overall"),
+        general_risk_constraint("hazard_exposure"),
+        general_risk_constraint("vulnerability"),
+        general_risk_constraint("coping_capacity"),
+        CheckConstraint(
+            "(global_rank >= 1) AND (global_rank <= 250)", name="global_rank"
+        ),
         CheckConstraint(
             "meta_avg_recentness_years >= 0.0",
             name="meta_avg_recentness_years",
         ),
         CheckConstraint(
-            "(reference_period_end >= reference_period_start) OR (reference_period_start IS NULL)",
-            name="reference_period",
+            "(meta_missing_indicators_pct >= 0.0) AND (meta_missing_indicators_pct <= 100.0)",
+            name="meta_missing_indicators_pct",
         ),
+        reference_period_constraint(),
     )
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    resource_hdx_id: Mapped[int] = mapped_column(
+    resource_hdx_id: Mapped[str] = mapped_column(
         ForeignKey("resource.hdx_id", onupdate="CASCADE", ondelete="CASCADE"),
         nullable=False,
     )
     admin2_ref: Mapped[int] = mapped_column(
-        ForeignKey("admin2.id", onupdate="CASCADE"), nullable=False
+        ForeignKey("admin2.id", onupdate="CASCADE"), primary_key=True
     )
-    risk_class: Mapped[int] = mapped_column(Integer, nullable=False)
+    risk_class: Mapped[RiskClass] = mapped_column(
+        Enum(RiskClass), nullable=False
+    )
     global_rank: Mapped[int] = mapped_column(Integer, nullable=False)
     overall_risk: Mapped[float] = mapped_column(Float, nullable=False)
     hazard_exposure_risk: Mapped[float] = mapped_column(Float, nullable=False)
@@ -59,12 +73,11 @@ class DBNationalRisk(Base):
         Float, nullable=True
     )
     reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, nullable=True, server_default=text("NULL"), index=True
+        DateTime, primary_key=True
     )
     reference_period_end: Mapped[datetime] = mapped_column(
-        DateTime, nullable=True, server_default=text("NULL"), index=True
+        DateTime, nullable=False, index=True
     )
-    source_data: Mapped[str] = mapped_column(Text, nullable=True)
 
     resource = relationship("DBResource")
     admin2 = relationship("DBAdmin2")
@@ -75,15 +88,6 @@ view_params_national_risk = ViewParams(
     metadata=Base.metadata,
     selectable=select(
         *DBNationalRisk.__table__.columns,
-        DBDataset.hdx_id.label("dataset_hdx_id"),
-        DBDataset.hdx_stub.label("dataset_hdx_stub"),
-        DBDataset.title.label("dataset_title"),
-        DBDataset.hdx_provider_stub.label("dataset_hdx_provider_stub"),
-        DBDataset.hdx_provider_name.label("dataset_hdx_provider_name"),
-        DBResource.name.label("resource_name"),
-        DBResource.update_date.label("resource_update_date"),
-        DBResource.hapi_updated_date.label("hapi_updated_date"),
-        DBResource.hapi_replaced_date.label("hapi_replaced_date"),
         DBLocation.code.label("location_code"),
         DBLocation.name.label("location_name"),
         DBAdmin1.code.label("admin1_code"),