+++ tests/conftest.py
@@ -1,23 +1,26 @@
 from typing import List
 
 import pytest
-from hdx.database.views import build_view
-from sqlalchemy import create_engine, insert, text
+from hdx.database import Database
+from sqlalchemy import (
+    insert,
+)
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.declarative import DeclarativeMeta
-from sqlalchemy.orm import sessionmaker
 
 from hapi_schema.db_admin1 import DBAdmin1
 from hapi_schema.db_admin2 import DBAdmin2
-from hapi_schema.db_age_range import DBAgeRange
 from hapi_schema.db_dataset import DBDataset
-from hapi_schema.db_food_security import DBFoodSecurity
-from hapi_schema.db_gender import DBGender
-from hapi_schema.db_humanitarian_needs import DBHumanitarianNeeds
-from hapi_schema.db_ipc_phase import DBIpcPhase
-from hapi_schema.db_ipc_type import DBIpcType
+from hapi_schema.db_food_security import (
+    DBFoodSecurity,
+)
+from hapi_schema.db_humanitarian_needs import (
+    DBHumanitarianNeeds,
+)
 from hapi_schema.db_location import DBLocation
-from hapi_schema.db_national_risk import DBNationalRisk
+from hapi_schema.db_national_risk import (
+    DBNationalRisk,
+)
 from hapi_schema.db_operational_presence import (
     DBOperationalPresence,
 )
@@ -25,20 +28,14 @@ from hapi_schema.db_org import DBOrg
 from hapi_schema.db_org_type import DBOrgType
 from hapi_schema.db_patch import DBPatch
 from hapi_schema.db_population import DBPopulation
-from hapi_schema.db_population_group import DBPopulationGroup
-from hapi_schema.db_population_status import DBPopulationStatus
 from hapi_schema.db_resource import DBResource
 from hapi_schema.db_sector import DBSector
-from hapi_schema.utils.base import Base
+from hapi_schema.views import prepare_hapi_views
 from sample_data.data_admin1 import data_admin1
 from sample_data.data_admin2 import data_admin2
-from sample_data.data_age_range import data_age_range
 from sample_data.data_dataset import data_dataset
 from sample_data.data_food_security import data_food_security
-from sample_data.data_gender import data_gender
 from sample_data.data_humanitarian_needs import data_humanitarian_needs
-from sample_data.data_ipc_phase import data_ipc_phase
-from sample_data.data_ipc_type import data_ipc_type
 from sample_data.data_location import data_location
 from sample_data.data_national_risk import data_national_risk
 from sample_data.data_operational_presence import data_operational_presence
@@ -46,23 +43,18 @@ from sample_data.data_org import data_org
 from sample_data.data_org_type import data_org_type
 from sample_data.data_patch import data_patch
 from sample_data.data_population import data_population
-from sample_data.data_population_group import data_population_group
-from sample_data.data_population_status import data_population_status
 from sample_data.data_resource import data_resource
 from sample_data.data_sector import data_sector
 
 
 @pytest.fixture(scope="session")
-def engine():
-    engine = create_engine(url="sqlite:///:memory:")
-
-    # Execute pragma statement to enable foreign key constraints
-    with engine.connect() as conn:
-        conn.execute(text("PRAGMA foreign_keys = ON;"))
-
+def session():
     # Build the DB
-    Base.metadata.create_all(engine)
-    session = sessionmaker(bind=engine)()
+    db_uri = "postgresql+psycopg://postgres:postgres@localhost:5432/hapitest"
+    database = Database(
+        db_uri=db_uri, recreate_schema=True, prepare_fn=prepare_hapi_views
+    )
+    session = database.get_session()
 
     # Populate all tables
     session.execute(insert(DBDataset), data_dataset)
@@ -72,15 +64,9 @@ def engine():
     session.execute(insert(DBAdmin1), data_admin1)
     session.execute(insert(DBAdmin2), data_admin2)
 
-    session.execute(insert(DBPopulationStatus), data_population_status)
-    session.execute(insert(DBPopulationGroup), data_population_group)
     session.execute(insert(DBOrgType), data_org_type)
     session.execute(insert(DBOrg), data_org)
     session.execute(insert(DBSector), data_sector)
-    session.execute(insert(DBIpcPhase), data_ipc_phase)
-    session.execute(insert(DBIpcType), data_ipc_type)
-    session.execute(insert(DBGender), data_gender)
-    session.execute(insert(DBAgeRange), data_age_range)
 
     session.execute(insert(DBNationalRisk), data_national_risk)
     session.execute(insert(DBPopulation), data_population)
@@ -91,35 +77,31 @@ def engine():
     session.execute(insert(DBPatch), data_patch)
 
     session.commit()
-
-    return engine
+    return session
 
 
 @pytest.fixture(scope="session")
-def run_view_test(engine):
+def run_view_test(session):
     def _run_view_test(view, whereclause):
-        Base.metadata.create_all(engine)
         select_instance = view.select().where(*whereclause)
-        select_instance.compile(bind=engine)
-        result = engine.connect().execute(select_instance)
+        result = session.execute(select_instance)
         assert result.fetchone()
 
     return _run_view_test
 
 
 @pytest.fixture(scope="session")
-def run_constraints_test(engine):
+def run_constraints_test(session):
     def _run_constraints_test(
         new_rows: List[DeclarativeMeta], expected_constraint: str
     ):
         """Test that a constraint will be triggered by passing its name
         and a list of one or more rows that violate it."""
-        Base.metadata.create_all(engine)
-        session = sessionmaker(bind=engine)()
         for new_row in new_rows:
             session.add(new_row)
         with pytest.raises(IntegrityError) as exc_info:
             session.commit()
+        session.rollback()
         assert expected_constraint in str(exc_info.value)
 
     return _run_constraints_test