+++ tests/sample_data/data_location.py
@@ -2,12 +2,9 @@ from datetime import datetime
 
 data_location = [
     dict(
-        id=1,
         code="FOO",
         name="Foolandia",
         reference_period_start=datetime(2023, 1, 1),
         reference_period_end=None,
-        hapi_updated_date=datetime(2023, 1, 1),
-        hapi_replaced_date=None,
     )
 ]