+++ tests/test_org.py
@@ -1,22 +1,18 @@
-from datetime import datetime
+from hdx.database import Database
 
-from hdx.database.views import build_view
-
-from hapi_schema.db_org import DBOrg, view_params_org
+from hapi_schema.db_org import view_params_org
 
 
 def test_org_view(run_view_test):
     """Check that org view references org type."""
-    view_org = build_view(view_params_org.__dict__)
+    view_org = Database.prepare_view(view_params_org.__dict__)
     run_view_test(
         view=view_org,
         whereclause=(
-            view_org.c.id == 1,
             view_org.c.org_type_code == "433",
             view_org.c.org_type_description == "Donor",
         ),
     )
-
 
 def test_org_vat(run_indexes_test, run_columns_test):
     """Check that the org view as table is correct - columns match, expected indexes present"""