+++ tests/test_sector.py
@@ -1,11 +1,11 @@
-from hdx.database.views import build_view
+from hdx.database import Database
 
 from hapi_schema.db_sector import view_params_sector
 
 
 def test_sector_view(run_view_test):
     """Check that sector view shows some columns."""
-    view_sector = build_view(view_params_sector.__dict__)
+    view_sector = Database.prepare_view(view_params_sector.__dict__)
     run_view_test(
         view=view_sector,
         whereclause=(